fix : UCS-4 character evaluation and unify string runtime functions

+++ src/libasr/codegen/llvm_utils.h
@@ -499,7 +499,7 @@ class ASRToLLVMVisitor;
              * Allocate heap memory for string.
              * Notice : It doesn't set the length.
             */
-            void set_string_memory_on_heap(ASR::string_physical_typeType str_physical_type, llvm::Value* str, llvm::Value* len, int char_kind = 1);
+            void set_string_memory_on_heap(ASR::string_physical_typeType str_physical_type, llvm::Value* str, llvm::Value* len, int64_t char_kind = 1);
 
             /*
              * Allocate stack memory for string.
